rename packages + main func
